+++ src/game/initRacing.js
@@ -1,0 +1,7 @@
+export function initRacing(round, carNames) {
+  return {
+    round,
+    carNames,
+    carDistances: Array(carNames.length).fill(0),
+  };
+}
