+++ src/result/printRacing.js
@@ -1,10 +1,10 @@
 import { Console } from "@woowacourse/mission-utils";
 
-export function printRacing(carCount, carNames, carDistances) {
+export function printRacing(carNames, carDistances) {
   Console.print("실행 결과");
   const DASH = "-";
 
-  for (let i = 0; i < carCount; i++) {
+  for (let i = 0; i < carDistances.length; i++) {
     Console.print(`${carNames[i]} : ${DASH.repeat(carDistances[i])}`);
   }
 }
